Add copy feature for code blocks and title bar
In order to make copying code from the article really easy, this commit
adds a copy button to every code block in the article. When hovering a
codeblock, the copy button appears and allows grabbing the codeblock's
contents. Additionally, this commit adds macos style title bars to every
codeblock to make the codeblocks more visually appealing.